implement user tracking backend application
This application is responsible for receiving requests from the tracker javascript library hook.

It uses Sidekiq along with Redis to enqueue page hits and leads creation, in order to prevent server and database congestion, and even crashes.

Its main focus is request speed, this way somethings as success responses are ignored due to async processing, so the application would need to halt while trying to retrieve a response, and this doesn't make much of a difference when gathering leads.